+++ environment.yml
@@ -24,3 +24,4 @@ dependencies:
   - pip
   - pip:
     - tensorflow
+    - torch
